% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combDunnett.R
\name{combDunnett}
\alias{combDunnett}
\title{Calculate the inverse normal combination z statistic based on Dunnett's test}
\usage{
combDunnett(w, z, d)
}
\arguments{
\item{w}{A vector of pre-specified weights for incremental test statistics up to the current stage.}

\item{z}{A matrix of cumulative log-rank test statistics across stages (up to current stage) for individual hypotheses that are in the intersection hypothesis. Columns represent dose levels, rows represent stages.}

\item{d}{A matrix of cumulative number of events across stages (up to current stage) for dose levels that are included in the intersection hypothesis. Columns represent dose levels, rows represent stages.}
}
\value{
The inverse normal Z statistic for the Dunnett's test
}
\description{
For an intersection hypothesis, calculate the inverse normal combination test based on Dunnett's test up to current stage for cohort 1.
}
\examples{
# Global intersection H with 3 dose arms. Perform testing at stage 3 post dose selection
w <- c(sqrt(2/3),sqrt(1/6),sqrt(1/6))
z <- matrix(c(1.1, 0.8,1.2,
             1.5, 1.3,1.3,
             1.8, 1.6,1.6),nrow = 3, ncol = 3, byrow = TRUE)
d <- matrix(c(20, 18,19,
             22, 21,23,
             25, 23,23),nrow = 3, ncol = 3, byrow = TRUE)
combDunnett(w,z,d)

}
