% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combTest.R
\name{combTest}
\alias{combTest}
\title{Calculate inverse normal combination z statistics combining cohort 1 and 2}
\usage{
combTest(w, h, z1, d1, z2, sid, zbound)
}
\arguments{
\item{w}{A vector of pre-specified weights for incremental test statistics across stages based on cohort 1. Same for all intersection hypotheses.}

\item{h}{A vector of pre-specified weights for cohort 1 and cohort 2, respectively, at current stage. Same for all intersection hypotheses.}

\item{z1}{A matrix of cumulative log-rank test statistics across stages (up to current stage) for individual hypotheses that are in the intersection hypothesis. Columns represent dose levels, rows represent stages.}

\item{d1}{A matrix of cumulative number of events across stages (up to current stage) for dose levels that are included in the intersection hypothesis. Columns represent dose levels, rows represent stages.}

\item{z2}{The log-rank test statistic at current stage for the selected dose arm based on cohort 2.}

\item{sid}{ID of the selected dose}

\item{zbound}{Standardized efficacy boundary}
}
\value{
The inverse normal combination test Z statistics for all possible intersection hypotheses.
}
\description{
For the selected dose, calculate inverse normal combination z statistics combining cohort 1 and 2 at current stage for all possible intersection hypotheses that include the selected dose
}
\examples{
# One dose is selected from 3 dose arms, and performing testing at stage 3 post dose selection
# calculate boundary at 1st stage
w <- c(sqrt(2/3),sqrt(1/6),sqrt(1/6))
h <- c(sqrt(60/700),sqrt(1-60/700))
z1 <- matrix(c(0.7,  1,   0.7,
              0.8,  1.1, 0.9,
              0.85, 1.2, 1),nrow = 3, ncol = 3, byrow = TRUE)
d1 <- matrix(c(20, 18,19,
              22, 21,23,
              25, 23,23),nrow = 3, ncol = 3, byrow = TRUE)
z2 <- 2.1
sid <- 2 # dose 2 is selected
zbound <- 2.01
combTest(w,h,z1,d1, z2, sid,zbound)

}
